Move base server into common
